Refactor readme and controller
