+++ scene/realistic-lighting-and-shadows/src/main/resources/realistic_lighting_and_shadows/main.fxml
@@ -24,7 +24,7 @@
 <StackPane fx:controller="com.esri.samples.realistic_lighting_and_shadows.RealisticLightingAndShadowsController"
            xmlns="http://javafx.com/javafx"
            xmlns:fx="http://javafx.com/fxml" stylesheets="/realistic_lighting_and_shadows/style.css">
-    <!--SDK SceneView-->
+    <!--SceneView-->
     <SceneView fx:id="sceneView"/>
     <StackPane.margin>
         <Insets topRightBottomLeft="10"/>
@@ -34,6 +34,6 @@
         <Label fx:id="timeLabel"/>
         <ComboBox fx:id="comboBox" maxWidth="Infinity">
         </ComboBox>
-        <Slider fx:id="timeSlider" max="1440" value="720" onMouseDragged="#updateTimeOfDay"/>
+        <Slider fx:id="timeSlider" showTickLabels="true" showTickMarks="true" majorTickUnit="240" minorTickCount="60" max="1440" value="720" onMouseDragged="#updateTimeOfDay"/>
     </VBox>
 </StackPane>
